Django_Application/ml_app/utils/face_extraction.py
import os
from dataclasses import dataclass
from typing import List, Optional, Tuple

import cv2
import numpy as np
from PIL import Image


@dataclass
class FaceExtractionResult:
    faces: List[Image.Image]  # PIL RGB face crops
    boxed_frames: List[np.ndarray]  # BGR frames with bounding boxes drawn
    num_faces: int


def _evenly_spaced_indices(frame_count: int, n: int) -> List[int]:
    if frame_count <= 0:
        return [0] * n
    if n <= 1:
        return [max(frame_count // 2, 0)]
    return np.linspace(0, frame_count - 1, num=n, dtype=int).tolist()


def _ensure_dir(path: str) -> None:
    os.makedirs(path, exist_ok=True)


def extract_faces_from_video(
    video_path: str,
    sequence_length: int,
    *,
    allow_fallback_center_crop: bool = True,
) -> FaceExtractionResult:
    """
    Extract up to 1 face per sampled frame using MTCNN when available.
    Falls back to face_recognition if facenet-pytorch is not installed.

    Returns:
      - faces: list of PIL RGB crops (length <= sequence_length)
      - boxed_frames: sampled frames with bounding boxes drawn (length <= sequence_length)
      - num_faces: number of frames where a face was detected
    """
    sequence_length = int(sequence_length)
    if sequence_length <= 0:
        raise ValueError("sequence_length must be > 0")

    cap = cv2.VideoCapture(video_path)
    if not cap.isOpened():
        raise RuntimeError(f"Unable to open video: {video_path}")

    frame_count = int(cap.get(cv2.CAP_PROP_FRAME_COUNT) or 0)
    indices = _evenly_spaced_indices(frame_count, sequence_length)

    # Prefer MTCNN (stronger detection); fallback to face_recognition.
    mtcnn = None
    try:
        from facenet_pytorch import MTCNN  # type: ignore

        mtcnn = MTCNN(keep_all=False, device="cpu")
    except Exception:
        mtcnn = None

    fr = None
    if mtcnn is None:
        try:
            import face_recognition  # type: ignore

            fr = face_recognition
        except Exception:
            fr = None

    faces: List[Image.Image] = []
    boxed_frames: List[np.ndarray] = []
    num_faces = 0

    def center_crop_as_face(frame_rgb: np.ndarray) -> Tuple[Image.Image, Tuple[int, int, int, int]]:
        h, w = frame_rgb.shape[:2]
        side = int(min(h, w) * 0.6)
        side = max(side, 64)
        cx, cy = w // 2, h // 2
        x1 = max(0, cx - side // 2)
        y1 = max(0, cy - side // 2)
        x2 = min(w, x1 + side)
        y2 = min(h, y1 + side)
        crop = frame_rgb[y1:y2, x1:x2]
        return Image.fromarray(crop), (x1, y1, x2, y2)

    for idx in indices:
        cap.set(cv2.CAP_PROP_POS_FRAMES, int(idx))
        ok, frame_bgr = cap.read()
        if not ok or frame_bgr is None:
            continue

        # Convert for detectors
        frame_rgb = cv2.cvtColor(frame_bgr, cv2.COLOR_BGR2RGB)

        box: Optional[Tuple[int, int, int, int]] = None  # (x1,y1,x2,y2)
        if mtcnn is not None:
            # MTCNN expects PIL or numpy RGB
            boxes, _ = mtcnn.detect(frame_rgb)
            if boxes is not None and len(boxes) > 0:
                x1, y1, x2, y2 = boxes[0].tolist()
                box = (int(x1), int(y1), int(x2), int(y2))
        elif fr is not None:
            locs = fr.face_locations(frame_rgb)
            if locs:
                top, right, bottom, left = locs[0]
                box = (left, top, right, bottom)

        if box is None:
            if allow_fallback_center_crop:
                # Demo-friendly fallback: keeps UI working when detectors aren't available.
                face_pil, fallback_box = center_crop_as_face(frame_rgb)
                faces.append(face_pil)
                num_faces += 1

                boxed = frame_bgr.copy()
                x1, y1, x2, y2 = fallback_box
                cv2.rectangle(boxed, (x1, y1), (x2, y2), (30, 64, 175), 2)
                boxed_frames.append(boxed)
            else:
                # Enterprise/real inference: if no face is detected, do not fabricate crops.
                boxed_frames.append(frame_bgr)
            continue

        x1, y1, x2, y2 = box
        x1 = max(0, x1)
        y1 = max(0, y1)
        x2 = min(frame_rgb.shape[1], x2)
        y2 = min(frame_rgb.shape[0], y2)
        if x2 <= x1 or y2 <= y1:
            boxed_frames.append(frame_bgr)
            continue

        num_faces += 1

        # Crop face
        crop_rgb = frame_rgb[y1:y2, x1:x2]
        face_pil = Image.fromarray(crop_rgb)
        faces.append(face_pil)

        # Draw box for UI
        boxed = frame_bgr.copy()
        cv2.rectangle(boxed, (x1, y1), (x2, y2), (30, 64, 175), 2)  # electric blue
        boxed_frames.append(boxed)

    cap.release()
    return FaceExtractionResult(faces=faces, boxed_frames=boxed_frames, num_faces=num_faces)

